Add files via upload
This code generates a grid of rows and columns specified by the user, randomly fills it with either "S" or "O", and then counts the number of occurrences of the "SOS" sequence. The program repeats this process 99 times and then calculates the total number of "SOS" sequences found and their average. The input validation is also implemented to ensure that the user inputs integers for the number of rows and columns.